requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
requests>=2.31.0          # HTTP client for Binance REST API
python-dotenv>=1.0.0      # Load .env into os.environ

# ── CLI ────────────────────────────────────────────────────────────────────────
typer[all]>=0.12.0        # Typer CLI framework (includes rich for colours)
rich>=13.7.0              # Terminal formatting

# ── Web UI (optional bonus) ────────────────────────────────────────────────────
streamlit>=1.35.0         # Lightweight web UI
pandas>=2.0.0             # Used in the history table in UI

# ── Networking / reliability ───────────────────────────────────────────────────
urllib3>=2.0.0            # Retry logic (via requests adapter)

# ── Testing ────────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-cov>=5.0.0         # Coverage reporting

# ── Dev / linting (optional) ──────────────────────────────────────────────────
# flake8>=7.0.0
# mypy>=1.10.0
